# docs/swagger.yaml
definitions:
  entities.Case:
    properties:
      case_description:
        type: string
      case_name:
        type: string
      created_at:
        type: string
      datums:
        items:
          $ref: '#/definitions/entities.Datum'
        type: array
      drill_depth:
        type: number
      fluids:
        items:
          $ref: '#/definitions/entities.Fluid'
        type: array
      holes:
        items:
          $ref: '#/definitions/entities.Hole'
        type: array
      id:
        type: string
      pipe_size:
        type: number
      strings:
        items:
          $ref: '#/definitions/entities.String'
        type: array
    type: object
...
  entities.Datum:
    properties:
      air_gap:
        type: number
      created_at:
        type: string
      datum_description:
        type: string
      datum_elevation:
        type: number
      ground_elevation:
        type: number
      id:
        type: string
      name:
        type: string
      system_description:
        type: string
      system_elevation:
        type: number
      type:
        type: string
      wellhead_elevation:
        type: number
    type: object
  entities.Design:
    properties:
      actual_date:
        type: string
      created_at:
        type: string
      id:
        type: string
      plan_name:
        type: string
      stage:
        type: string
      trajectories:
        items:
          $ref: '#/definitions/entities.Trajectory'
        type: array
      version:
        type: string
    type: object
...
  entities.Fluid:
    properties:
      base_fluid_id:
        type: string
      case_id:
        type: string
      created_at:
        type: string
      density:
        type: number
      description:
        type: string
      fluid_base_type_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  entities.FrictionFactor:
    properties:
      back_reaming_casing:
        type: number
      back_reaming_open_hole:
        type: number
      created_at:
        type: string
      hole_id:
        type: string
      id:
        type: string
      rotating_off_bottom_casing:
        type: number
      rotating_off_bottom_open_hole:
        type: number
      rotating_on_bottom_casing:
        type: number
      rotating_on_bottom_open_hole:
        type: number
      slide_drilling_casing:
        type: number
      slide_drilling_open_hole:
        type: number
      tripping_in_casing:
        type: number
      tripping_in_open_hole:
        type: number
      tripping_out_casing:
        type: number
      tripping_out_open_hole:
        type: number
    type: object
  entities.Hole:
    properties:
      case_id:
        type: string
      casings:
        items:
          $ref: '#/definitions/entities.HoleCasing'
        type: array
      created_at:
        type: string
      friction_factors:
        items:
          $ref: '#/definitions/entities.FrictionFactor'
        type: array
      id:
        type: string
      open_holes:
        items:
          $ref: '#/definitions/entities.OpenHole'
        type: array
    type: object
  entities.HoleCasing:
    properties:
      burst_rating:
        type: number
      collapse_rating:
        type: number
      created_at:
        type: string
      description:
        type: string
      drift_id:
        type: number
      effective_hole_diameter:
        type: number
      friction_factor:
        type: number
      grade:
        type: string
      hole_id:
        type: string
      id:
        type: string
      id_value:
        description: To avoid collision with ID field
        type: number
      length:
        type: number
      linear_capacity:
        type: number
      manufacturer:
        type: string
      md_base:
        type: number
      md_top:
        type: number
      min_yield_strength:
        type: number
      model:
        type: string
      od:
        type: number
      shoe_md:
        type: number
      weight:
        type: number
    type: object
  entities.OpenHole:
    properties:
      created_at:
        type: string
      description:
        type: string
      effective_diameter:
        type: number
      friction_factor:
        type: number
      hole_id:
        type: string
      id:
        type: string
      id_value:
        description: To avoid collision with ID field
        type: number
      length:
        type: number
      linear_capacity:
        type: number
      md_base:
        type: number
      md_top:
        type: number
      volume_excess:
        type: number
    type: object
...
  entities.Section:
    properties:
      created_at:
        type: string
      id:
        type: string
      section_type_id:
        type: string
      string_id:
        type: string
      values:
        items:
          $ref: '#/definitions/entities.SectionValue'
        type: array
    type: object
  entities.SectionValue:
    properties:
      attribute_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      section_id:
        type: string
      value:
        type: string
    type: object
...
  entities.String:
    properties:
      case_id:
        type: string
      created_at:
        type: string
      depth:
        type: number
      id:
        type: string
      name:
        type: string
      sections:
        items:
          $ref: '#/definitions/entities.Section'
        type: array
    type: object
  entities.Trajectory:
    properties:
      cases:
        items:
          $ref: '#/definitions/entities.Case'
        type: array
      created_at:
        type: string
      description:
        type: string
      headers:
        items:
          $ref: '#/definitions/entities.TrajectoryHeader'
        type: array
      id:
        type: string
      name:
        type: string
      units:
        items:
          $ref: '#/definitions/entities.TrajectoryUnit'
        type: array
    type: object
  entities.TrajectoryHeader:
    properties:
      created_at:
        type: string
      customer:
        type: string
      field:
        type: string
      id:
        type: string
      job_number:
        type: string
      kelly_bushing_elev:
        type: number
      profile:
        type: string
      profile_type:
        type: string
      project:
        type: string
      structure:
        type: string
      wellhead:
        type: string
      your_ref:
        type: string
    type: object
  entities.TrajectoryUnit:
    properties:
      azim:
        type: number
      created_at:
        type: string
      dogleg:
        type: number
      global_e_coord:
        type: number
      global_n_coord:
        type: number
      id:
        type: string
      incl:
        type: number
      local_e_coord:
        type: number
      local_n_coord:
        type: number
      md:
        type: number
      sub_sea:
        type: number
      tvd:
        type: number
      vertical_section:
        type: number
    type: object
...
  entities.Well:
    properties:
      active_well_unit:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        type: string
      name:
        type: string
      type:
        type: string
      universal_well_identifier:
        type: string
      well_number:
        type: string
      wellbores:
        items:
          $ref: '#/definitions/entities.Wellbore'
        type: array
      working_group:
        type: string
    type: object
  entities.Wellbore:
    properties:
      average_column_rotation_frequency:
        type: number
      average_hook_load:
        type: number
      average_inlet_flow:
        type: number
      average_torque:
        type: number
      average_weight_on_bit:
        type: number
      bottom_hole_location:
        type: string
      created_at:
        type: string
      depth_interval:
        type: number
      designs:
        items:
          $ref: '#/definitions/entities.Design'
        type: array
      down_static_friction:
        type: number
      id:
        type: string
      maximum_column_rotation_frequency:
        type: number
      maximum_torque:
        type: number
      maximum_weight_on_bit:
        type: number
      name:
        type: string
      riser_pressure:
        type: number
      wellbore_depth:
        type: number
    type: object
  helpers.Response:
    properties:
      message:
        type: string
    type: object
...
  requests.CreateWellRequest:
    properties:
      body:
        $ref: '#/definitions/requests.CreateWellRequestBody'
      siteID:
        type: string
    type: object
  requests.CreateWellRequestBody:
    properties:
      active_well_unit:
        type: string
      description:
        type: string
      location:
        type: string
      name:
        type: string
      type:
        type: string
      universal_well_identifier:
        type: string
      well_number:
        type: string
      working_group:
        type: string
    type: object
...
  requests.UpdateWellRequest:
    properties:
      body:
        $ref: '#/definitions/requests.UpdateWellRequestBody'
      id:
        type: string
    type: object
  requests.UpdateWellRequestBody:
    properties:
      active_well_unit:
        type: string
      description:
        type: string
      location:
        type: string
      name:
        type: string
      type:
        type: string
      universal_well_identifier:
        type: string
      well_number:
        type: string
      working_group:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/wells:
    get:
      consumes:
      - application/json
      description: Retrieves all wells
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Site ID
        in: query
        name: siteId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Well'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Get Wells
      tags:
      - wells
    post:
      consumes:
      - application/json
      description: Creates a new well
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Site ID
        in: query
        name: siteId
        required: true
        type: string
      - description: Well input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CreateWellRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Create Well
      tags:
      - wells
  /api/v1/wells/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing well
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Well ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Delete Well
      tags:
      - wells
    get:
      consumes:
      - application/json
      description: Retrieves a well by its ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Well ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Well'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Get Well by ID
      tags:
      - wells
    put:
      consumes:
      - application/json
      description: Updates an existing well
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Well ID
        in: path
        name: id
        required: true
        type: string
      - description: Well input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateWellRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Well'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Update Well
      tags:
      - wells
swagger: "2.0"